Add WaNet / Input-aware experiment scripts and library-compat patches
Patches (24 files) — adapt BackdoorBench to NumPy 2.0 / PyTorch 2.6+ / grad-cam 1.5+:
- utils/trainer_cls.py: np.infty -> np.inf, torch.load weights_only=False
- utils/save_load_attack.py: torch.load weights_only=False
- analysis/visual_gradcam.py: drop use_cuda arg from FullGrad
- defense/{abl,ac,anp,bnp,clp,d-br,d-st,dbd,ep,fp,ft,ft-sam,i-bau,mcr,nab,nad,nc,npd,rnp,sau,spectral}.py:
  --num_workers type=float -> type=int (fixes DataLoader worker spawn)

New analysis scripts (7 files):
- analysis/compare_wanet_clean_vs_bd.py: 3-row (clean / bd / cross) visual diff
- analysis/plot_wanet_defense_compare.py: ACC/ASR/RA grouped bar chart across defenses
- analysis/plot_anp_threshold_sweep.py: ANP mask threshold Pareto frontier
- analysis/plot_layer_weight_delta.py: per-stage |W_attack - W_defense| ratios + top-20 params
- analysis/plot_s_ablation.py: WaNet warping-strength s ablation (3-panel)
- analysis/md_to_html.py: markdown -> HTML renderer with Korean-friendly CSS
- sh_run_s_ablation.sh: WaNet s in {0.1,0.3,0.5,0.7,1.0} sweep runner

+++ defense/d-br.py
@@ -295,7 +295,7 @@ class d_br(defense):
     
         parser.add_argument('--epochs', type=int)
         parser.add_argument('--batch_size', type=int)
-        parser.add_argument("--num_workers", type=float)
+        parser.add_argument("--num_workers", type=int)
         parser.add_argument('--lr', type=float)
         parser.add_argument('--lr_scheduler', type=str, help='the scheduler of lr')
         parser.add_argument('--steplr_stepsize', type=int)
