+++ defense/fp.py
@@ -121,7 +121,7 @@ class FinePrune(defense):
 
         parser.add_argument('--epochs', type=int)
         parser.add_argument('--batch_size', type=int)
-        parser.add_argument("--num_workers", type=float)
+        parser.add_argument("--num_workers", type=int)
         parser.add_argument('--lr', type=float)
         parser.add_argument('--lr_scheduler', type=str, help='the scheduler of lr')
